Annotate holding registers with units, general cleanup
Note that the manual and Modbus disagree on temperature units.  The manual
says temperatures have a unit of 0.5degC, while the number Modbus returns is
in tenths of a degree (ie value = reading*0.1).  The manual appears to mean
resolution (ie to nearest half a degree) rather than the unit of the number
returned.