Clean up, add units, add guessed slots for error codes
See previous commit for note that units represent the scale factor for the
figures on Modbus, not the resolution of the reading as given in the service
manual.

Error code slots assigned because there's a gap in the table that matches
the service manual, and the values don't look like normal 16 bit signed
integers (or floats or strings).